feature(implement): add company signup functionality [Finishes #168132789]
