+++ src/services/index.js
@@ -1,5 +1,8 @@
 import UserService from './UserService';
 import BookingService from './BookingService';
 import CompanyService from './companyService';
+import SupplierService from './supplierService';
 
-export { UserService, BookingService, CompanyService };
+export {
+  UserService, BookingService, CompanyService, SupplierService
+};
