+++ src/validation/userValidation.js
@@ -24,15 +24,15 @@ const complexityOptions = {
    * 1) signup - validates user upon registration.
    * 2) dummy - a dummy method for testing validations upon success.
    */
-export default class authValidation {
-/**
+export default class userValidation {
+  /**
      * Validates user paramenters upon registration
      *
      * @param {object} userObject - The user object
      * @param {object} res - The user response object
      * @returns {object} - returns an object (error or response).
      */
-  static async userSignup(userObject) {
+  static async signup(userObject) {
     // joi parameters to test against user inputs
     const schema = {
       firstName: joi.string().min(3).max(25).required()
@@ -60,7 +60,7 @@ export default class authValidation {
       companyName: joi.string().min(3).max(40).required()
         .label('Please add your company name'),
     };
-      // Once user inputs are validated, move into server
+    // Once user inputs are validated, move into server
     const { error } = joi.validate({ ...userObject }, schema);
     if (error) {
       // throw errorResponse(res, { code: 400, message: error.details[0].context.label });
@@ -70,46 +70,12 @@ export default class authValidation {
   }
 
   /**
-       * Validates supplier paramenters upon registration
-       *
-       * @param {object} supplierObject - The supplier object
-       * @param {object} res - The supplier response object
-       * @returns {object} - returns an object (error or response).
-       */
-  static async supplierSignup(supplierObject) {
-    const schema = {
-      firstName: joi.string().min(3).max(25).required()
-        .label('Please enter a valid firstname \n the field must not be empty and it must be more than 2 letters'),
-      lastName: joi.string().min(3).max(25).required()
-        .label('Please enter a valid lastname \n the field must not be empty and it must be more than 2 letters'),
-      email: joi.string().email().required()
-        .label('Please enter a valid company email address'),
-      password: new passwordComplexity(complexityOptions).required()
-        .label('Password is required. \n It should be more than 8 characters, and should include at least a capital letter, and a number'),
-      phoneNumber: joi.string().regex(/^[0-9+\(\)#\.\s\/ext-]+$/).required()
-        .label('Please input a valid phone number'),
-      companyName: joi.string().min(3).max(40).required()
-        .label('Please add your company name'),
-      companyAddress: joi.string().min(3).max(40).required()
-        .label('Please add your company address'),
-      categoryOfServiceId: joi.valid([1, 2]).required()
-        .label('Please enter a valid categoryOfServiceId')
-    };
-    const { error } = joi.validate({ ...supplierObject }, schema);
-    if (error) {
-      throw error;
-    }
-    return true;
-  }
-
-
-  /**
      * Validates user paramenters upon registration
      *
      * @param {object} userObject - The user object
      * @param {object} res - The user response object
      * @returns {object} - returns an object (error or response).
-    */
+     */
   static async companySignup(userObject) {
     // joi parameters to test against user inputs
     const schema = {
@@ -140,6 +106,28 @@ export default class authValidation {
     return true;
   }
 
+  /**
+     * Validates user paramenters upon registration
+     *
+     * @param {object} userObject - The user object
+     * @param {object} res - The user response object
+     * @returns {object} - returns an object (error or response).
+     */
+  static async userLogin(userObject) {
+    // joi parameters to test against user inputs
+    const schema = {
+      email: joi.string().email().required()
+        .label('Please enter a valid company email address'),
+      password: new passwordComplexity(complexityOptions).required()
+        .label('Password is required. \n It should be more than 8 characters, and should include at least a capital letter, and a number'),
+    };
+    // Once user inputs are validated, move into server
+    const { error } = joi.validate({ ...userObject }, schema);
+    if (error) {
+      throw error;
+    }
+    return true;
+  }
 
   /**
    *  Dummy callback function for validation tests
